First draft of WP8/WP8.1 (desktop IE10, mobile IE10) support. Includes fix for emulated icon handling.
Capture pointer events by default so we get the same behaviour as IOS.
We call setPointerCapture on the element that receives the pointer down
event. Added FlexPropertyHelper enhancements to support IE10

do not blurBeforeAnimation if the node is the body element since this
causes the IE10 browser to disappear behind other windows

on going fixes. For ie10 use -ms-flex: 1 1 instead of -ms-flex: 1 since
ie10 sets shrink to 0 by default if not specified whereas webkit sets it
to 1. Do not do event.preventDefault in the CellList if we are running
on windows phone, prevents the scroll panel the CellList is in scrolling
properly.

for ie10/ie11 desktop needed to add a focus fix where when using pointer
capture, input elements failed to get the focus properly. See
https://stackoverflow.com/questions/27355271/desktop-ie10-ie11-on-windows-input-elements-fail-to-get-focus-when-pointer-cap

capturing pointer events on input or textarea elements causes many
issues, so we no longer capture pointer events if the target element is
an input or textarea. Turned on default behaviour for textarea so we can
scroll textarea. However this can cause a bounce when scrolling a text
area to the start or end. However this is better then not being able to
scroll at all like currently in IOS

remove drop down arrow on select

fixed issue where scroll panel did not work on windows phone 8.1 update
but worked fine on windows phone 8. Had to force getComputedStyle to
return a 3D matrix for the transform property by specifying a non-zero z
value (-1px). We now detect the device density correctly for ie10/11 (we
use the screen object). We do not try and detect ie11 user agent, we
assume if using ie11 that the meta tag ie10 compatibilty is set. User
agent strings for ie11 are now even more complex. They specify webkit,
iphone, gecko etc etc as part of the string. This makes OS detection
more difficult using the user agent string.

added flexwrap to FlexInputHelper, corrected name of MozAlignSelf in
Mozilla implementation of FlexInputHelper

Need access to ButtonBaseAppearance so subclass can override the default
touch handlers added that do not always give you the required behaviour

Added correct orientation support for windows phone 8.1

In IEOrientationHandler removed window.alert left in by mistake. Added
!important to portraitonly and landscapeonly css since not always easy
to ensure this css gets priority over other css

Avoid long emulation in ScrollPanel

fix minor bug when specifying ie10 compatibility. Does not seem to make
any diferrence. Looks like you have to specify in the html file as a
meta tag and not rely on it being dynamically added and picked up by
ie10/11

due to poor performance of ScrollPanel in IE10/11 when a reasonable
amount of DOM added, attempting to eliminate any translate3d type
issues. Use 0px instead of -1px for the z co-ordinate and check what
transform matrix returned, either 2d or 3D and extract the relevant
co-ordinates

removed -ms-touch-action: none, which is applied to elements since need
to use native scrollable divs to get better performance then using the
MGWT ScrollPanel. Unfortunately one side effect of this is we start to
get application bounce when scrolling limits are reached of a scrollbale
div.

Only apply ios71 body bug fix when ios71. Also stop possible re-creation
of orientation handler if current orientation handler not removed
correctly. Remove the orientation handler correctly when orientation
event fired.

Added more position information to the TapEvent via the Touch object.
When preventing scrolling for IOS we no longer preventDefault the
TouchMove event if it originates from an INPUT, TEXTAREA or SELECT
element. This solves a multitude of issues with these input types where
default behaviour of the browser is required so that they function as
expected. Also on more sensitive touch devices like the ipad air 2 the
input type elements were very hard to tap and navigate between if a
TouchMove event was generated as part of the touch and the TouchMove
vent is prevent defaulted.

Updated to use GWT 2.7

ImageConverter modified to wait for the image to be loaded before trying
to convert it

Fixes #220.

Revert "Updated to use GWT 2.7"

This reverts commit 073e6a0c61ba67b31d29759903c3123f0d5102bf.