Fix for black background gradient not being applied on Android.
The linear gradient is never applied on Chrome 39 both on the desktop and
Android browser when mgwt.os is set to 'android'. The result is the background
stays the default #ECEBF1 color with white text making it difficult to read.

I'm not sure if the body::before pseudo element is still needed for legacy
reasons, however, the submitted change does work in Chrome 39.